requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-admin-tools==0.9.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
Pillow==8.1.0
psycopg2==2.8.6
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
